refactor: move algorithms to folder
