+++ src/test/kotlin/modelTests/algo/LouvainTests.kt
@@ -1,4 +1,4 @@
-package modelTests
+package modelTests.algo
 
 import graph.Graph
 import graph.Vertex
