+++ src/test/kotlin/modelTests/algo/SCCTests.kt
@@ -1,4 +1,4 @@
-package modelTests
+package modelTests.algo
 
 import graph.Graph
 import graph.Vertex
